Adding a view extension and using it to centralize visibility logic and to save some lines of code in the future.
Updating the target sdkVersion to 30, so android 11 users can also use the app. And small change in CaptureImagePresenterTest.kt

I wanted to deleted JCenter because it's going to be closed soon:
https://jfrog.com/blog/into-the-sunset-bintray-jcenter-gocenter-and-chartcenter/
So I checked which libraries are not in maven central yet, and I  noticed the conditionwatcher and the scabbard library. I noticed that conditionwatcher isn't maintained anymore so I created a module with the arr, but in case of scabbard I saw that the creator is doing the migration and it will be done in the 0.5.0 version, so at the end I decided to leave that library and not create the aar and also leave jcenter and delete the jcenter when scabbard is updated.